fix: backend review findings - TokenService, ContextController, PendingMessageProcessor
TokenService:
- Fixed clamp comment to accurately describe purpose (ceil overflow)
- Fixed savingsPercent type: explicit (double) cast + 0.0 default

ContextController:
- Removed redundant equals check in isWithinProject() (startsWith already covers it)

PendingMessageProcessor:
- Added AtomicBoolean overlap guard to prevent concurrent @Scheduled executions
- Uses compareAndSet + finally reset pattern

Also committed cron patrol updates:
- backend-review-findings.md: new review #5 findings
- patrol-log.md: updated patrol log entries

+++ backend/src/main/java/com/ablueforce/cortexce/service/TokenService.java
@@ -47,7 +47,7 @@ public class TokenService {
             // Fallback to empty string if serialization fails
             size += 2; // "[]"
         }
-        // Clamp to Integer.MAX_VALUE to prevent overflow (use long literal to avoid overflow)
+        // Clamp to Integer.MAX_VALUE: ceil() can exceed int range for very large observations
         size = Math.min(size, 2L * Integer.MAX_VALUE);
         return (int) Math.ceil(size / CHARS_PER_TOKEN);
     }
@@ -66,8 +66,8 @@ public class TokenService {
 
         int savings = totalDiscoveryTokens - totalReadTokens;
         double savingsPercent = totalDiscoveryTokens > 0
-            ? Math.round((double) savings / totalDiscoveryTokens * 100)
-            : 0;
+            ? (double) Math.round((double) savings / totalDiscoveryTokens * 100)
+            : 0.0;
 
         return new TokenEconomics(
             observations.size(),
